Refactor: move outbox functionality to service
