+++ lotti/lib/blocs/journal/persistence_cubit.dart
@@ -6,7 +6,6 @@ import 'package:bloc/bloc.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_native_timezone/flutter_native_timezone.dart';
 import 'package:lotti/blocs/journal/persistence_state.dart';
-import 'package:lotti/blocs/sync/outbox_cubit.dart';
 import 'package:lotti/classes/audio_note.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/classes/entry_links.dart';
@@ -23,25 +22,23 @@ import 'package:lotti/location.dart';
 import 'package:lotti/main.dart';
 import 'package:lotti/services/notification_service.dart';
 import 'package:lotti/services/vector_clock_service.dart';
+import 'package:lotti/sync/outbox.dart';
 import 'package:lotti/sync/vector_clock.dart';
 import 'package:lotti/utils/file_utils.dart';
 import 'package:lotti/utils/timezone.dart';
 import 'package:uuid/uuid.dart';
 
 class PersistenceCubit extends Cubit<PersistenceState> {
-  late final OutboxCubit _outboundQueueCubit;
   final JournalDb _journalDb = getIt<JournalDb>();
   late final VectorClockService _vectorClockService;
   final InsightsDb _insightsDb = getIt<InsightsDb>();
+  final OutboxService _outboxService = getIt<OutboxService>();
 
   final uuid = const Uuid();
   DeviceLocation? location;
   Timer? timer;
 
-  PersistenceCubit({
-    required OutboxCubit outboundQueueCubit,
-  }) : super(PersistenceState.initial()) {
-    _outboundQueueCubit = outboundQueueCubit;
+  PersistenceCubit() : super(PersistenceState.initial()) {
     _vectorClockService = getIt<VectorClockService>();
     init();
   }
@@ -386,7 +383,7 @@ class PersistenceCubit extends Cubit<PersistenceState> {
       await _journalDb.addTagged(journalEntity);
 
       if (saved && enqueueSync) {
-        await _outboundQueueCubit.enqueueMessage(SyncMessage.journalEntity(
+        await _outboxService.enqueueMessage(SyncMessage.journalEntity(
           journalEntity: journalEntity,
           status: SyncEntryStatus.initial,
         ));
@@ -405,7 +402,7 @@ class PersistenceCubit extends Cubit<PersistenceState> {
 
         await _journalDb.upsertEntryLink(link);
 
-        await _outboundQueueCubit.enqueueMessage(SyncMessage.entryLink(
+        await _outboxService.enqueueMessage(SyncMessage.entryLink(
           entryLink: link,
           status: SyncEntryStatus.initial,
         ));
@@ -644,10 +641,12 @@ class PersistenceCubit extends Cubit<PersistenceState> {
       await saveJournalEntityJson(journalEntity);
 
       if (enqueueSync) {
-        await _outboundQueueCubit.enqueueMessage(SyncMessage.journalEntity(
-          journalEntity: journalEntity,
-          status: SyncEntryStatus.update,
-        ));
+        await _outboxService.enqueueMessage(
+          SyncMessage.journalEntity(
+            journalEntity: journalEntity,
+            status: SyncEntryStatus.update,
+          ),
+        );
       }
       await transaction.finish();
 
@@ -663,7 +662,7 @@ class PersistenceCubit extends Cubit<PersistenceState> {
   Future<int> upsertEntityDefinition(EntityDefinition entityDefinition) async {
     int linesAffected =
         await _journalDb.upsertEntityDefinition(entityDefinition);
-    await _outboundQueueCubit.enqueueMessage(SyncMessage.entityDefinition(
+    await _outboxService.enqueueMessage(SyncMessage.entityDefinition(
       entityDefinition: entityDefinition,
       status: SyncEntryStatus.update,
     ));
@@ -672,7 +671,7 @@ class PersistenceCubit extends Cubit<PersistenceState> {
 
   Future<int> upsertTagEntity(TagEntity tagEntity) async {
     int linesAffected = await _journalDb.upsertTagEntity(tagEntity);
-    await _outboundQueueCubit.enqueueMessage(SyncMessage.tagEntity(
+    await _outboxService.enqueueMessage(SyncMessage.tagEntity(
       tagEntity: tagEntity,
       status: SyncEntryStatus.update,
     ));
